requirements.txt
Django==1.5.8
django-appconf==0.6
-e git://github.com/jezdez/django_compressor.git@772ecd1ef2ce021d05cbc44eb8602f7d59db2c52#egg=django_compressor
django-easy-pdf==0.1.0
django-filebrowser==3.5.2
django-grappelli==2.4.10
django-ipware==1.1.6
django-mptt==0.6.0
django-picklefield==0.3.0
django-tinymce==1.5.3
django-widget-tweaks==1.3
FeinCMS==1.7.4
html5lib==1.0b8
MySQL-python==1.2.5
Pillow==3.4.2
PyPDF2==1.26.0
python-dateutil==2.1
pytils==0.2.3
pytz===2013d
reportlab==3.3.0
selenium==3.0.2
six==1.10.0
sorl-thumbnail==12.3
South==1.0.2
webencodings==0.5
xhtml2pdf==0.0.6
XlsxWriter==0.9.6
